+++ src/model/application/modal_interaction.rs
@@ -61,7 +61,7 @@ pub struct ModalInteraction {
     /// For monetized applications, any entitlements of the invoking user.
     pub entitlements: Vec<Entitlement>,
     /// Attachment size limit in bytes
-    pub attachment_size_limit: u64,
+    pub attachment_size_limit: u32,
 }
 
 #[cfg(feature = "model")]
